+++ src/lib.rs
@@ -109,14 +109,15 @@ pub use dts::{DtsAudioMode, DtsBitRate, DtsFrameType, DtsHeader, DtsSampleRate, 
 pub use error::{Error, Result};
 pub use ifo::{
     menu_existence, AudioApplicationMode, AudioAttributes, AudioCodingMode, AudioLanguageType,
-    AudioQuantizationDrc, CellAddrEntry, CellBlockType, CellCategory, CellPlaybackInfo,
-    CellPositionInfo, CellType, DvdChapter, DvdTitle, DvdTitleEntry, FrameRate, McExtensionEntry,
-    MenuAttributes, MenuType, NavCommand, PaletteEntry, Pgc, PgcCommandTable, PgcTime, Pgci,
-    PgciLu, PgciLuSrp, PgciSrp, PgciUt, PgciUtSrp, PtlMait, Ptt, PttTitle, SubpictureAttributes,
-    SubpictureCodingMode, SubpictureLanguageType, TitleAttributes, TmapEntry, TtSrpt,
-    VideoAspectRatio, VideoAttributes, VideoCodingMode, VideoResolution, VideoStandard, VmgIfo,
-    VmgPtlMait, VmgVtsAtrt, VmgVtsAtrtEntry, VobuAdmap, VtsCAdt, VtsIfo, VtsPttSrpt, VtsTmap,
-    VtsTmapti, VtsiMat, DVD_SECTOR, VMG_MAGIC, VTS_MAGIC,
+    AudioQuantizationDrc, AudioStreamControl, CellAddrEntry, CellBlockType, CellCategory,
+    CellPlaybackInfo, CellPositionInfo, CellType, DvdChapter, DvdTitle, DvdTitleEntry, FrameRate,
+    McExtensionEntry, MenuAttributes, MenuType, NavCommand, PaletteEntry, Pgc, PgcCommandTable,
+    PgcTime, Pgci, PgciLu, PgciLuSrp, PgciSrp, PgciUt, PgciUtSrp, PlaybackMode, PtlMait, Ptt,
+    PttTitle, SubpictureAttributes, SubpictureCodingMode, SubpictureLanguageType,
+    SubpictureStreamControl, TitleAttributes, TmapEntry, TtSrpt, VideoAspectRatio, VideoAttributes,
+    VideoCodingMode, VideoResolution, VideoStandard, VmgIfo, VmgPtlMait, VmgVtsAtrt,
+    VmgVtsAtrtEntry, VobuAdmap, VtsCAdt, VtsIfo, VtsPttSrpt, VtsTmap, VtsTmapti, VtsiMat,
+    DVD_SECTOR, VMG_MAGIC, VTS_MAGIC,
 };
 pub use iso9660::{
     DirectoryRecord, Iso9660Entry, Iso9660Volume, PathTableEntry, PrimaryVolumeDescriptor,